refactory : 기능 함수 최소 단위 모듈화 진행 (버튼 제외)
