+++ frontend/src/components/Home/CurrentTimeText.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Text } from 'react-native'
+import { roundStyles } from '../../styles/roundStyles'
+
+const CurrentTimeText = ({ currentTime }) => {
+  const formattedTime = currentTime.toLocaleTimeString()
+  return <Text style={roundStyles.timeText}>{formattedTime}</Text>
+}
+
+export default CurrentTimeText
