+++ frontend/src/components/Home/TimelineCircle.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Animated } from 'react-native'
-import { timelineStyles } from '../../styles/HomeStyles'
+import { timelineStyles } from '../../styles/timelineStyles'
 
 const TimelineCircle = ({ stopStyle }) => {
   return <Animated.View style={[timelineStyles.circle, stopStyle]} />
